refactor: moved payments-related method on dedicated service
