+++ src/main/java/it/gov/pagopa/fdr/controller/OrganizationsController.java
@@ -5,6 +5,7 @@ import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
 import it.gov.pagopa.fdr.controller.model.payment.response.PaginatedPaymentsResponse;
 import it.gov.pagopa.fdr.service.FlowService;
+import it.gov.pagopa.fdr.service.PaymentService;
 import it.gov.pagopa.fdr.service.model.FindFlowsByFiltersArgs;
 import it.gov.pagopa.fdr.service.re.model.FdrActionEnum;
 import it.gov.pagopa.fdr.util.Re;
@@ -14,9 +15,12 @@ public class OrganizationsController implements IOrganizationsController {
 
   private final FlowService flowService;
 
-  protected OrganizationsController(FlowService flowService) {
+  private final PaymentService paymentService;
+
+  protected OrganizationsController(FlowService flowService, PaymentService paymentService) {
 
     this.flowService = flowService;
+    this.paymentService = paymentService;
   }
 
   @Override
@@ -58,7 +62,7 @@ public class OrganizationsController implements IOrganizationsController {
       long pageNumber,
       long pageSize) {
 
-    return this.flowService.getPaymentsFromPublishedFlow(
+    return this.paymentService.getPaymentsFromPublishedFlow(
         FindFlowsByFiltersArgs.builder()
             .organizationId(organizationId)
             .pspId(pspId)
