+++ src/main/java/it/gov/pagopa/fdr/service/FlowService.java
@@ -3,18 +3,12 @@ package it.gov.pagopa.fdr.service;
 import it.gov.pagopa.fdr.Config;
 import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
-import it.gov.pagopa.fdr.controller.model.payment.response.PaginatedPaymentsResponse;
 import it.gov.pagopa.fdr.exception.AppErrorCodeMessageEnum;
 import it.gov.pagopa.fdr.exception.AppException;
 import it.gov.pagopa.fdr.repository.FdrFlowRepository;
-import it.gov.pagopa.fdr.repository.FdrPaymentRepository;
 import it.gov.pagopa.fdr.repository.entity.common.RepositoryPagedResult;
 import it.gov.pagopa.fdr.repository.entity.flow.FdrFlowEntity;
-import it.gov.pagopa.fdr.repository.entity.flow.projection.FdrFlowIdProjection;
-import it.gov.pagopa.fdr.repository.entity.payment.FdrPaymentEntity;
-import it.gov.pagopa.fdr.repository.enums.FlowStatusEnum;
 import it.gov.pagopa.fdr.service.middleware.mapper.FlowMapper;
-import it.gov.pagopa.fdr.service.middleware.mapper.PaymentMapper;
 import it.gov.pagopa.fdr.service.middleware.validator.SemanticValidator;
 import it.gov.pagopa.fdr.service.model.FindFlowsByFiltersArgs;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -30,26 +24,15 @@ public class FlowService {
 
   private final FdrFlowRepository flowRepository;
 
-  private final FdrPaymentRepository paymentRepository;
-
   private final FlowMapper flowMapper;
 
-  private final PaymentMapper paymentMapper;
-
   public FlowService(
-      Logger log,
-      Config cachedConfig,
-      FdrFlowRepository flowRepository,
-      FdrPaymentRepository paymentRepository,
-      FlowMapper flowMapper,
-      PaymentMapper paymentMapper) {
+      Logger log, Config cachedConfig, FdrFlowRepository flowRepository, FlowMapper flowMapper) {
 
     this.log = log;
     this.cachedConfig = cachedConfig;
     this.flowRepository = flowRepository;
-    this.paymentRepository = paymentRepository;
     this.flowMapper = flowMapper;
-    this.paymentMapper = paymentMapper;
   }
 
   public PaginatedFlowsResponse getPaginatedPublishedFlows(FindFlowsByFiltersArgs args) {
@@ -117,44 +100,5 @@ public class FlowService {
 
     log.debugf("Entity found. Mapping data to final response.");
     return this.flowMapper.toSingleFlowResponse(result);
-  }
-
-  public PaginatedPaymentsResponse getPaymentsFromPublishedFlow(FindFlowsByFiltersArgs args) {
-
-    /*
-    MDC.put(EVENT_CATEGORY, EventTypeEnum.INTERNAL.name());
-    String action = (String) MDC.get(ACTION);
-    MDC.put(ORGANIZATION_ID, organizationId);
-    MDC.put(FDR, fdr);
-    MDC.put(PSP_ID, psp);
-     */
-
-    String organizationId = args.getOrganizationId();
-    String pspId = args.getPspId();
-    String flowName = args.getFlowName();
-    long revision = args.getRevision();
-    long pageNumber = args.getPageNumber();
-    long pageSize = args.getPageSize();
-
-    log.debugf(
-        "Executing query on payments related on flow by organizationId [%s], pspId [%s] flowName"
-            + " [%s], revision:[%s]",
-        organizationId, pspId, flowName, revision);
-
-    ConfigDataV1 configData = cachedConfig.getClonedCache();
-    SemanticValidator.validateGetPaymentsFromPublishedFlow(configData, args);
-
-    FdrFlowIdProjection flowIdProjection =
-        this.flowRepository.findIdByOrganizationIdAndPspIdAndName(
-            organizationId, pspId, flowName, revision, FlowStatusEnum.PUBLISHED);
-    if (flowIdProjection == null) {
-      throw new AppException(AppErrorCodeMessageEnum.REPORTING_FLOW_NOT_FOUND, flowName);
-    }
-
-    RepositoryPagedResult<FdrPaymentEntity> paginatedResult =
-        this.paymentRepository.findByFlowObjectId(
-            flowIdProjection.getId(), (int) pageNumber, (int) pageSize);
-
-    return paymentMapper.toPaginatedPaymentsResponse(paginatedResult, pageSize, pageNumber);
   }
 }
